# Add performance monitoring (CPU and memory usage) functionality to Timer